Rename IAST fixture file
