+++ tests/appsec/iast/sink/test_xpath_injection.py
@@ -3,7 +3,7 @@
 # Copyright 2021 Datadog, Inc.
 
 from utils import features
-from .._test_iast_fixtures import BaseSinkTestWithoutTelemetry
+from ..utils import BaseSinkTestWithoutTelemetry
 
 
 @features.iast_sink_xpathinjection
